pyproject.toml
[project]
name = "fk-project"
version = "0.1.0"
description = "books.toscrape.com product monitor"
authors = [
    {name = "toludaree", email = "isaactoluwani30@gmail.com"},
]
dependencies = ["pydantic>=2.12.4", "beautifulsoup4>=4.14.2", "httpx[brotli,zstd]>=0.28.1", "python-dotenv>=1.2.1", "pymongo>=4.15.3", "apscheduler>=3.11.1", "fastapi[standard]>=0.122.0", "passlib>=1.7.4", "bcrypt==4.3.0", "python-jose[cryptography]>=3.5.0"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm-backend"
